Over for now
I set myself a visual presentation too big for me to hadle right now and it's keeping me from progressing in my study. I may come back to it and make a pixel perfect presentation of it in the future, but for now this project is done. The fundamentals are here, in register you can add to a fake database with users and their passwords whilst in login you authenticate them.

See you in the next one!